support spark map_from_arrays semantics
