+++ velox/functions/sparksql/CMakeLists.txt
@@ -27,6 +27,7 @@ velox_add_library(
   LeastGreatest.cpp
   MakeTimestamp.cpp
   Map.cpp
+  MapFromArrays.cpp
   RegexFunctions.cpp
   Size.cpp
   String.cpp
